move LocalStorage tests to a separate file
